requirements.txt
playwright>=1.40.0
aiogram>=3.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
openai>=1.68.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
firecrawl-py>=2.0.0
